i18n: Change all FFMD web addresses to HTTPS
We previously linked to http://register.md.freifunk.net/ which is not
optimal, because users enter personal data and/or some login token.
After fixing the SSL certificate on the server, we can now link to
https://register.md.freifunk.net/

Note: This is the same change as already proposed in pull requests #79
and #80 which are superseded by this. However after merging
5983856bd27333549e2dc5bed2f104f4ebdecfe9 with pull request #93 we could
not easily merge those old changes based on an earlier version.
Reimplementing those was trivial however, so I just did that instead of
rebasing, adapting, and squashing.

Suggested-by: Michel Vorsprach <michel@vorsprach.de>